Move validation functions and tests to a separate go file
