+++ api/v1alpha1/validation.go
@@ -1,0 +1,66 @@
+package v1alpha1
+
+import (
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/validation/field"
+	"sigs.k8s.io/controller-runtime/pkg/webhook"
+)
+
+var _ webhook.Validator = &EtcdPeer{}
+
+// ValidateCreate validates that all required fields are present and valid.
+func (o *EtcdPeer) ValidateCreate() error {
+	path := field.NewPath("spec")
+	var allErrs field.ErrorList
+
+	allErrs = append(
+		allErrs,
+		o.Spec.Storage.validate(path.Child("storage"))...,
+	)
+	return allErrs.ToAggregate()
+}
+
+// ValidateCreate validates that deletion is allowed
+// TODO: Not yet implemented
+func (o *EtcdPeer) ValidateDelete() error {
+	var allErrs field.ErrorList
+	return allErrs.ToAggregate()
+}
+
+// ValidateCreate validates that only supported fields are changed
+// TODO: Not yet implemented
+func (o *EtcdPeer) ValidateUpdate(old runtime.Object) error {
+	var allErrs field.ErrorList
+	return allErrs.ToAggregate()
+}
+
+func validatePersistentVolumeClaimSpec(path *field.Path, o *corev1.PersistentVolumeClaimSpec) field.ErrorList {
+	var allErrs field.ErrorList
+	if o == nil {
+		allErrs = append(allErrs, field.Required(path, ""))
+		return allErrs
+	}
+	if o.StorageClassName == nil {
+		allErrs = append(allErrs, field.Required(path.Child("storageClassName"), ""))
+		return allErrs
+	}
+	if _, ok := o.Resources.Requests["storage"]; !ok {
+		allErrs = append(allErrs, field.Required(path.Child("resources", "requests", "storage"), ""))
+		return allErrs
+	}
+	return allErrs
+}
+
+func (o *EtcdPeerStorage) validate(path *field.Path) field.ErrorList {
+	var allErrs field.ErrorList
+	if o == nil {
+		allErrs = append(allErrs, field.Required(path, ""))
+		return allErrs
+	}
+	allErrs = append(
+		allErrs,
+		validatePersistentVolumeClaimSpec(path.Child("volumeClaimTemplate"), o.VolumeClaimTemplate)...,
+	)
+	return allErrs
+}
